+++ src/player/sleep_timer/__init__.py
@@ -1,0 +1,6 @@
+from .dialog import SLEEP_TIMER_DIALOG_TITLE, SleepTimerDialog
+
+__all__ = [
+    "SLEEP_TIMER_DIALOG_TITLE",
+    "SleepTimerDialog",
+]